+++ src/main/java/com/nylas/JobStatusQuery.java
@@ -1,0 +1,5 @@
+package com.nylas;
+
+public class JobStatusQuery extends RestfulQuery<JobStatusQuery> {
+
+}